requirements.txt
# requirements.txt
requests==2.26.0
pycryptodome==3.11.0
pytz==2021.3
pycryptodomex==3.11.0
pycryptodome==3.11.0
cryptography==36.0.0
urllib3==1.26.8
